requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
psycopg2-binary
flask-login